requirements.txt
telethon>=1.34.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
pandas>=2.1.0
openpyxl>=3.1.2

